+++ Tests/DangerTests/BitBucketServerTestResources/BitBucketServerUser.swift
@@ -1,0 +1,6 @@
+
+public let BitBucketServerUserJSON = """
+
+    {"name":"tum-id","emailAddress":"user.name@tum.de","id":1,"displayName":"Firstname Lastname","active":true,"slug":"tum-id","type":"NORMAL","links":{"self":[{"href":"https://repobruegge.in.tum.de/users/tum-id"}]}}
+
+"""